Add factorial and TailCall to the set.
Made a simple TailCall interface and used it calculate the factorial for
any given number.
Uses Tail Call Recursion (with the help of the TailCall checked in) to
avoid having a StackPointer Overflow.
Made two classes, one that uses simple ints (/Integer) and one that uses
BigInteger. Things get big fast. And so I made a helper that decides
if the int or the BigInteger way should be used.
I've done no tests, but I would like to think that it would be faster
to use the primitive types instead of the BigInteger objects. Hence
the use of the primitives until they are too small.